+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.1
 
 - ⭐ Neu: Nutze `require("ak.template.eep-web-main")` in einer Anlage ohne eigenen Lua-Code um EEP-Web zu verwenden
+- ⭐ Neu: Nutze `AkStartWithDebug=true` wenn Du ausführliche Informationen im Log sehen möchtest
 
 ## 0.10.0